+++ src/app/editor/component_registry.rs
@@ -21,6 +21,9 @@ impl App {
         self.register_component("YSort", |world, e| {
             world.add_component(e, crate::ysort::YSort::default());
         });
+        self.register_component("AnimationEvents", |world, e| {
+            world.add_component(e, crate::animation::events::AnimationEvents::default());
+        });
         self.register_component("ParticleEmitter", |world, e| {
             world.add_component(e, crate::particle::ParticleEmitter::default());
         });
@@ -66,6 +69,9 @@ impl App {
         });
         self.register_component_remover("YSort", |world, e| {
             world.remove_component::<crate::ysort::YSort>(e);
+        });
+        self.register_component_remover("AnimationEvents", |world, e| {
+            world.remove_component::<crate::animation::events::AnimationEvents>(e);
         });
         self.register_component_remover("ParticleEmitter", |world, e| {
             world.remove_component::<crate::particle::ParticleEmitter>(e);